Test more windows password encoding command output cases
Use the same TEST_ALL_CREDENTIALS technique as is used in
CredentialsTest.

If TEST_ALL_CREDENTIALS is true, then use all of the
string sources in the test, otherwise only use a random sample of
strings.

If TEST_ALL_CREDENTIALS is not set, then use all string
sources if not running in a Jenkins job.