Escape double quote in windows credential responses
While testing ssh passphrase based login from Windows, I found that the
passphrase needs to be escaped in the echo command line, just as the
username and password are escaped in the echo command line.

Special rules apply to arguments to echo which are not surrounded by
double quotes.  Those rules are referenced in the comment on the related
utility method in this class.