Optionally detach ssh authenticated git calls from terminal
When I run CredentialsTest from my terminal window on Ubuntu 16.04, the
test fails for ssh keys which need a passphrase.  If I run the tests
from my IDE, or if I prepend "setsid" to the maven command that runs
the tests, the tests pass.

The ssh command called by git seems to require the DISPLAY variable, and
the GIT_SSH variable, and must be detached from the controlling terminal.
If any one of those is missing (at least on Ubuntu 16), the passphrase
prompt will not be answered.

Command line maven builds now include the property:

  org.jenkinsci.plugins.gitclient.CliGitAPIImpl.useSETSID=true

Default runtime value of the property is false. Users will not run
with this change unless they specifically set that property to true.
Most users don't run Jenkins with a controlling terminal attached,
so they don't need the change.

Sets useSETSID=true in surefire target so that command line invocations
of the tests will prefix the "git" command with setsid when it is used
in an ssh private key context.

Allows command line run of CredentialsTest with passphrase protected
private keys.

Setting BatchMode=yes in the ssh command does not have the same result.
The setsid call was the only technique I found that reliably allowed
the ssh call performed by command line git to consistently process the
script defined in the SSH_ASKPASS variable.

See [JENKINS-20879] and [JENKINS-25194] for more details.