Test more credentials variations
Need an additional bug fix before the git.addCredentials() test can
be enabled.  When using a username and password embedded in an http or
https URL, the JGit implementation is able to use the credentials (from
the automated test) if they are added with git.addDefaultCredentials.
Unfortunately, JGit is not able to use the credential if it is added
with git.addCredentials.

It seems like those two things should be equivalent.  I suspect there
may be some difference due to the username / password embedded in the
test URL.