[JENKINS-43931] support private keys w/passphrase in Windows dirs with spaces
The temporary file protection code (correctly) prefers to place
sensitive temporary files near the workspace rather than placing them
in the system temporary directory.  The Windows git implementation
(through at least git 2.12.2) fails to authenticate if the value of
GIT_SSH is a path which includes a space.

If the workspace temporary directory name contains a space, the system
temporary directory will be used instead.  There is already code in
the CliGitAPIImpl class which writes a warning if the system temporary
directory includes a space in its path.