Fix special characters bugs in credentials workspace path
JENKINS-44041 - Windows authenticated git checkout fails if '(' or ')' in path to workspace
JENKINS-43931 - Windows authenticated git checkout fails if ' ' in path to workspace
JENKINS-44127 - Authenticated git checkout fails if '%' in path to workspace (Windows & Linux)

Also safeguards against use for "`" (grave) in a workspace path.
Jenkins already guards against that, but the added safety check is low
cost and passes the credentials tests.