Test JENKINS-40023 - JGit impl handles max changelog incorrectly
JGit changelog does not work well when setting max revisions limit,
if the revisions containt merge commits.  That is because RevWalker
will only return the number of revision as set by limitations, but
ChangelogCommand default behaviour is to skip merge commits. Resulting
in returning less commits then requested by the user.

CliGit changelog works correctly.

The test shows the failure so that if the bug is fixed, this test
will fail.