[JENKINS-45729] fix getRepository() to work with symref git files
The current implementation of getRepository() fails to work if the
repository directory in question has a .git file (known as a symref)
which points to the real directory. This causes problems if you try to
directly open a submodule or worktree repository using the git client.

This happens because FileRepositoryBuilder.create() uses .setGitDir() to
set the git directory which expects only a proper git directory path.
Instead, set the worktree and have the RepositoryBuilder figure out
where the git directory is.

This implementation is identical to the JGit implementation, which is
a nice bonus.

Additionally, add a test which shows the bug. Unfortunately, it does not
get run on the JGit implementation because of how the test submodules
are setup.

Without this fix, the git plugin will fail to fetch and checkout
revisions of git worktrees which use .git symref files, because it uses
the withRepository in order to obtain the git commit subject.