Narrow branch name matching test for fewer collisions
The previous test started failing when I submitted a pull request
from a branch that contained the letters "od".  That's too common
a character sequence to fail a test.

Ultimately this needs a better way of testing than using a local
mirror.  The local mirror creates a useful and interesting history,
but the same types of useful and interesting history could be
placed into a sample repository, or could be created by the test,
without the ongoing risk of a newly created branch causing a test
failure due to pattern matching.