Use structs 1.10 for clean transitive dependencies
The transitive dependencies in structs 1.9 caused the git plugin to need
an exclusion for its requireUpperBounds.  That exclusion was resolved by
updating the git plugin to depend on structs 1.10 instead of structs 1.9.

This change updates the git client plugin to structs 1.10 so that the
git plugin (and others) won't inadvertently have a transitive dependency
on structs 1.9 from the git client plugin.