Protect against ArrayIndexOutOfBoundsException
If "git branch -v --no-abbrev" output contains a line like " very-long-string-with-more-than-44-characters\n", it will bypass the first IF condition of the WHILE block and will end up being split. However, since it will only contain 1 token (as there are no spaces after applying "substring(2)"), accessing "branchVerboseOutput[1]" results in ArrayIndexOutOfBoundsException. This fix is to avoid such exception and pre-validate the array size (should be greater than 1 in order to contain, at least, the branch name and the SHA1).