Include the results of the command in the log when it fails
Proposed fix for JENKINS-47493.

Having the error in the exception is nice assuming someone can figure out where the exception renders, but most views of stage logs in Jenkins don't include the exception result at all. This patch causes it to log the stdout/stderr and status code to the log for the stage, so that it can be seen inline when it occurs.