Add [JENKINS-46054] failing tests
If the repository URL for a submodule includes ".url" in the
URL, then the submodule query code in the command line git
implementation gathers the wrong name as the name of the
submodule.

These tests check specific conditions with the regular
expression string used to parse the "git config"
output.

They intentionally fail with the current regular expression.