Add GitObject.equals and GitObject.hashCode
New GitClient.getTags() method returns a Set<GitObject>.

Testing the getTags() method is much easier if it implements equals.
If it implements equals, it needs to implement hashCode in order to
preserve the equals contract.

Uses EqualsVerifier to check the equals contract more thoroughly in
GitObjectTest and in the tests of its two subclasses, Tag and Branch.