Fail on findbugs warnings
Prior to requiring JDK at compile time and at run time, the Jenkins
versions being tested (1.625.3 and 2.60.1) were too different to have
a single code base be free of findbugs warnings with both Jenkins
versions.

Now that JDK 8 is required, the build can be configured to fail if a
new findbugs warning is introduced.