Fix both JENKINS-48818 and JENKINS-46054
Add JENKINS-48818 test - space character in submodule remote name

Submodule remote name which contains a space character was supported by
git client plugin versions prior to 2.7.0.  The bug fix for JENKINS-46054
(support submodule update when the remote name or the remote URL end with
'.url') broke support for submodule remote names which contain a space.

This includes a test to show that a remote name containing a space
character is not correctly handled by the 2.7.0 change.