Extra logging for [JENKINS-48357] diagnosis
I suspect that the git plugin is not the source of the problem that is
causing Pipeline jobs to hang.  Log the completion of a command line
git call in addition to the entry into a command line git call.

Do not merge!  This will dramatically increase the log spam from the
git plugin.  This is a pull request so that there is an easy way to
obtain a build of the plugin.