Assert annotated tag is null
The earlier assertNotEquals was not as strong as a specific assertion
that in the case of command line git, it is expected that an annotated
tag will not be found (and will thus return null).  It is a strange
condition that getHeadRev for an annotated tag returns null from
command line git and returns non-null from JGit, but that is the
reality.