Limit credentials test time rather than test count
The CredentialsTest uses a ClassRule JenkinsRule to reduce setup time
for each test.  A ClassRule JenkinsRule is safe because each of the
credentials tests is independent of all other credentials tests.

The JenkinsRule test is limited to not more than 180 seconds for a single
JenkinsRule test.  Because I have many authentication test configurations
(approaching 100 currently), the tests can fail because they reach the
JenkinsRule timeout.  Previously the failure was prevented by limiting
the number of tests which can be executed.

This change uses elapsed time as the limiter rather than number of tests.
All tests executed more than 130 seconds after the first CredentialsTest
will be skipped.  Since the order of CredentialsTests is intentionally
randomized and the tests are run on multiple platforms with multiple
Jenkins versions, this should give a reasonable chance that most
CredentialsTest combinations will be executed in a single CI run in my
home CI setup.

Note that the CredentialsTest depends on private credentials data.
The ci.jenkins.io execution will run at most one copy of this test if
it detects an RSA private key file on the agent executing the test.