Add JENKINS-50573 test case
When the fetch_() arguments included a URIish that is a remote name
(like "origin") instead of a remote URL (like
"git@github.com:jenkinsci/git-client-plugin.git"), the credential
detection code inside the CliGitAPIImpl did not realize that the
changes were required in the ssh based credentials.

GitSCMFileSystem passes a remote name as the argument instead of
passing the remote URL.  Using the remote name is elegant, since it
avoids a lookup of the remote name and relies on git to perform that
lookup itself.