Fix [JENKINS-50573] by passing remoteURL to launchCommandWithCredentials
CliGitAPIImpl.fetch_() takes an argument which includes a URIish that
generally represents the URL to the remote git repository.  However,
there are places in the git plugin (like GitSCMFileSystem) where the
caller uses the simplification of calling with the remote name ("origin")
instead of the remote URL.  That simplification avoids extracting or
remembering the remote URL for a repository which is already established.

Unfortunately, that simplification bypasses the code in CliGitAPIImpl
which inspects the URL being used and decides if authentication needs
to be adjusted to accomodate the incompatible change made in OpenSSH 7.7.

OpenSSH 7.7 has made a sensible choice.  The Jenkins git plugin relied
on an undocumented and unexpected behavior of OpenSSH versions prior to
OpenSSH 7.7.

Passing the remoteURL allows the CliGitAPIImpl code to adapt to the
OpenSSH 7.7 change.

Refer to the closed OpenSSH 7.7 bug report at
http://bugzilla.mindrot.org/show_bug.cgi?id=2849

Refer to the rejected OpenSSH pull request at
https://github.com/openssh/openssh-portable/pull/91

Refer to the closed Git for Windows issue at
https://github.com/git-for-windows/git/issues/1616