Remove ssh related comment from https username/password code segment
A comment about ssh authentication does not help the reader of the code
when it appears in a section which is using username/password instead
of private keys.

GIT_ASKPASS provides the path to the script which will be called to
answer git's prompt for credentials.  SSH_ASKPASS is likely not needed
in this section of the code, since the code is handling http and https,
not ssh private keys.