Move GitTool resolver test to separate file
The Windows tests depend on JenkinsRule being used as a ClassRule in
the GitToolTest.  Rather than sacrifice test execution time (which on
Windows is especially slow to start and stop JenkinsRule based tests),
move the shouldResolveToolsOnMaster to a separate test class that can
run its own JenkinsRule.

Extends the test to run on Windows in addition to Linux.