impl: fix @tmp directory creation
The workspace path variable is not holding the Jenkins jobs workspace
path in all cases. E.g. if 'checkout to subdirectory' is used it points
to a directory in the jobs workspace, resulting in @tmp directories in
it.

Using the WORKSPACE environment variable to create the @tmp directory
fixes this behavior.