JENKINS-55526 - reduce severity of reference repo warnings
The messages are only warnings, not errors.  Listing them as errors on the
listener is incorrect.  They are not fatal.  If a reference repository
is specified incorrectly, the message should alert the user because the
clone will need more disc space and more time.  The clone does not fail
when a reference repository is missing.