[JENKINS-55655] test default git client class type
JENKINS-55655 reports that Blue Ocean does not honor the proxy
configuration and other settings which are honored by other parts of
the Jenkins configuration.  The root of the problem seems to be that
Blue Ocean constructs a Git object with very reasonable defaults and
uses that to get a GitClient.  The resulting GitClient is a
JGitAPIImpl instead of a CliGitAPIImpl.

Since CLI git is the canonical implementation in the Jenkins git
client, the default should return a CLI git implementation, not a JGit
implementation.