Avoid test fails: 50 sec limit on credentials test
The junit test reporter plugin that reports test results during test
execution will incorrectly report a test failure if the unit test XML
file is kept open with incomplete content.  The credentials test leaves
the content open for the duration of the test and that frequently fails
due to the incomplete content in the open file.