Remove duplicated assertion
Duplication only detected on ancient git, but the code intended to move
the assertion into the conditional.

Failure detected with CentOS 6 and command line git 1.7.1.  Command line
git 1.7.1 is officially not supported by the git plugin, but there is
no reason to break users for those cases where it worked previously.