Fix JENKINS-56116 - Use raw format on ancient git
Specific format characters like '%B' are not supported on old git versions
like git 1.7.1 as included with CentOS 6.

Command line git 1.7.1 lacks several capabilities and is not officially
supported by the Jenkins git plugin.  However, some use cases work with
CentOS 6.  There is no need to break existing users for this specific
case, since it is easy enough to use the old 'raw' format as was done
in long ago versions of the git plugin.