Skip exception test if root user
The exception test depends on being unable to write a file to an
uncommon location.  If the tests are running as the root user, they
are allowed to write to that uncommon location.  If the tests are
running as an unprivileged user, they are not allowed to write to that
uncommon location.