Ignore getAllLogEntries test result
The getAllLogEntries method is deprecated.  Its test depends on precise
output ordering of the git command

  git log --all origin/master

The ordering is different between command line git and JGit when two
commits happen in the same second. JGit ordering is incorrect, since
it should be able to resolve the predecessor and successor
relationship from the relationship between the parent and child
commits.  However, that difference is consistently there in JGit 4.5
and in JGit 5.4.

The getAllLogEntries method in this API is deprecated and its output is
questionable, whether from JGit or command line git. Not valuable enough
to spend more time on it. I would have deleted the test completely but is
is easier and clearer to document when a test result is ignored rather
than to risk someone introducing a test again without knowing why the
output of the deprecated method is unreliable.