Better safeguard enabling repository sanity check
Summary: When the Jenkins command line argument disables repository sanity
checks, the security problem from SECURITY-1534 is visible in some cases.

SECURITY-1534 reports that repository URL's provided by the user were
not sanity checked before being passed to git ls-remote and git fetch. A
sanity check is now enabled by default.

As a backwards compatibility 'escape hatch', a Jenkins command line
argument can disable the repository sanity checks. Disabling the checks
then relies on command line git to perform the sanity checks.

GitClientSecurityTest uses a randomly selected value to enable or disable
the repository URL check. If the repository check is selected to be
disabled and the command line git implementation does not have full
support for the '--' separator between options and operands and the
attack string is one of a known set of strings, then the test forces
repository checks to be enabled so that it avoids cases where older
command line git doesn't sanity check its arguments.

If the test were run with repository checks disabled in those cases
on older command line git implementations (git 1.8.3 on CentOS 7,
git 2.7.4 on Ubuntu 16) it would cause the test code to not throw an
exception because those command line git versions do not fully support
'--' to separate options and operands.