Move SETSID control from pom to test
No need to complicate the pom file with setting a property, better to
adjust the value in the specific context of exactly the test cases
which need it.

The CALL_SETSID value only needs to be changed for those test cases
which are using a private key with a passphrase.  Private key tests
which don't use a passphrase and username/password tests do not need
to modify the CALL_SETSID value.