Fix testcase when cli git signs commit
org.jenkinsci.plugins.gitclient.GitAPITestCase#test_merge_with_message
assumed that the commit message was on line 7 of the revision output.
But when cli git is globally configured to sign each commit then line 7
is the start of the gpg signature (an empty line)

This fix checks if there is such a signature in the revision output
and instead assumes that the commit message is two lines below the
end of the signature.