Test with Jenkins 2.190.2 instead of 2.176.3
Do not bundle transitive dependency in jar.

The dependency on trilead-api should not bundle transitive dependencies
from the trilead-api in this jar.  Rely on the trilead-api plugin to
provide those dependencies.

Make trilead-api dependency optional.  Other trilead-api consumers
mentioned in JENKINS-59676 seem to all declare it as an optional
dependency.