Remove unused NotImplementedInJGit annotation
The NotImplementedInJGit annotation is only helpful when the test
runner uses it to decide if a test should be run.  That is a JUnit 3
technique. In JUnit 4, the Assume class is used to skip tests based on
runtime calculations.

Since the annotation was being ignored and the tests are passing, the
tests must be already coded to assert correctly for JGit, without
requiring that any of the tests are skipped.