Safely handle localMirror creation in tests
The localMirror() method clones the remote git client plugin repository
into a local copy for use by tests.  Using a remote clone is better
than cloning from the local copy because the local copy may be a shallow
clone or may be using a sparse checkout.

Unfortunately, the prior implementation of localMirror assumed exclusive
access to the target/clone.git directory.  When tests run in parallel,
that assumption of exclusive access is wrong.  Multiple tests would
attempt to populate the target/clone.git repository.  They would conflict
and fail the tests in strange and undesirable ways.

Safely populate the local mirror by cloning to a temporary directory then
moving the temporary directory to the intended final destination directory
after the clone is complete and after confirming that the destination
directory still does not exist.  There is still a brief window where
the directory is created after the existance check and before the atomic
move which renames the temporary directory to the correct final name.

The implementation accepts that if many unit tests were all executed at
the same time and were all using the localMirror, multiple clones will
happen and all but one of the clones will be discarded.  The tests reduce
the likelihood of that problem by adding a random delay to the start of
longer running tests that use the local mirror.  The delay was enough
in my tests to prevent the duplicate clones and the resulting discard
of the results of the clone.