Honor ibm.system.encoding property in passphrase file
Consnistent with usage in other locations in the plugin but requires
that a property be set on the agent JVM.  Setting a property on the
agent JVM may be a little more complicated than setting an environment
variable.  May require setting the argument in the advanced section of
the agent startup rather than in the agent properties.