JENKINS-60985 - Fix checkout serialization error
[JENKINS-60985](https://issues.jenkins-ci.org/browse/JENKINS-60985)
Fix checkout serialization error

The previous call to checkout was entirely inside the CliGitAPIImpl and
JGitAPIImpl classes.  The replacement used a CheckoutCommand that is
not serializable.  This change reverts the replacement.

Special thanks to Devin Nusbaum for identifying the issue.

Revert "Replace deprecated GitClient#checkout(...) calls"

This reverts a part of commit baec182192129519fa0639ba9771bd37be45d813.