Fix Windows test failure by enabling core.longpaths
One of the submodule cloning tests,
testSubmodulesUsedFromOtherBranches, fails on Windows agents when the
Windows workspace is in the default location on a Windows master.

Rather than fight with that directory location, enable core.longpaths
for the failing test.  Eventually we'll want to enable core.longpaths
on all command line git use from Windows computers.