Reduce test warning count by not using Assume
The maven surefire plugin reports skipped tests as a WARNING.
That WARNING causes the warnings plugin on ci.jenkins.io to count the
skipped test as a warning and display the warning in the job results.

The tests are intentionally skipped so they should not be counted as
warnings.  This converts the assertion from a warning to no warning.
The same code is executed as previously, without the maven surefire
warning.