Run JMH benchmarks only with benchmark.run=true
There is very little value to running the Java Microharness Benchmarks
to compare JGit and command line git because JGit changes very rarely.

A JMH test run without a review of the results is not very valuable.  Only
run the JMH tests if the property `benchmark.run` is `true`.

Always write JSON to the jmh-report.json file so that it can be
visualized with https://jmh.morethan.io/