Add hasGitRepo(boolean) to fix JENKINS-38996
The JGit implementation of hasGitRepo() always ignored parent
directories.  The CLI git impelementation of hasGitRepo() ignored parent
directories if the workspace did not contain a '.git' directory.
If the workspace contains an empty '.git' directory, then CLI git
will search parent directories.

That's a strange behavior, but I'd rather not learn by hard
experience which implementations depend on that strange behavior.

Introduces a new API, hasGitRepo(boolean checkParentDirectories).  The
JENKINS-38996 fix will require a new release of the git plugint that
depends on this release of the git client plugin and uses the new
API to avoid the poor behavior of the existing hasGitRepo() implementation.