Revert "More submodule test updates"
Plugin compatibility tester is an important part of the Jenkins plugin
BOM evaluation process.  The change in the `tests/getSubmodules` branch
breaks the submodule related tests on released plugin versions.  That
disrupts the plugin compatibility tester.

Return to the prior settings that were in the `tests/getSubmodules`
branch rather than disrupt PCT.

This reverts commit e208678d925f5bf13432f5ab963345ce782cf308.