Use container agent for CI build (#880)
* Use container agent for CI build

https://github.com/jenkinsci/bom/pull/1338#issuecomment-1198016897 notes
that there is an automated test failure in the bill of materials and
plugin compatibility tests.  This change is trying to duplicate that
failure on ci.jenkins.io.

* Exclude ssh tests on ci.jenkins.io Kubernetes agents

Unclear why the tests fail with a connection timeout on those agents and
only on those agents.  No issue on virtual machine based agents.  No
issue on static agents at my home.  No issue on any agents except the
agents on ci.jenkins.io.

* Exclude JSR-305 dependency from packaging

No need to bundle the JSR-305 transitive dependency from
spotbugs-annotation into the final hpi file.  The spotbugs annotations
are optional and the JSR-305 annotations should also be optional.
They have not been included in previous releases of the plugin.

* Test JDK 17 with Jenkins 2.346.1 LTS, not 2.342 weekly

* Do not fork tests on multi-core machines

* Remove unnecessary exclusion