Test a wider range of Jenkins versions in CI (#907)
* Test a wider range of Jenkins versions in CI

* Java 17 with the most recent LTS - 2.361.1
* Java 11 with the previous LTS - 2.346.3
* Java 8 with the default provided by jenkins.version - 2.332.4

* Avoid NPE in authenticated merge call

Not expected to occur in most production environments, but the spotbugs
warning with Jenkins 2.361.1 is correct.  If the repository URL is
null as read from the default remote, then a null could be passed to
launchCommandWithCredentials.  Better to safeguard against passing null
to a method that is annotated to not accept null in that parameter.