Use gson 2.8.9 to satisfy security scanners
The gson package is not enabled for serialization in Jenkins, so the
vulnerability reported as https://nvd.nist.gov/vuln/detail/CVE-2022-25647
is unlikely to have any impact on the git client plugin.  However, out
of an abundance of caution, it is better to upgrade the gson library
from 2.8.8 to 2.8.9.

https://github.com/google/gson/blob/master/CHANGELOG.md#version-289 is
the gson 2.8.9 changelog.