Allow file:// clones of submodules in tests
Allow local git clones to use the file:// protocol by setting
protocol.file.allow=always in the global git configuration for
the current user.

Command line git 2.38.1 and patches to earlier versions
disallow local git submodule cloning with the file:// protocol.
The change resolves a security issue but that security issue is
not a threat to these tests.

After the test completes, the setting is removed from the
global git configuration of the current user.

The setting must be applied to the global configuration of the
current user or the affected tests will fail. It is unfortunate
that the tests are temporarily modifying the configuration of the
current user, but it is more important that the tests be run
than that we defend the git settings of git plugin developers
from changes performed by the tests.