Allow automated tests from shallow clones
Do not attempt to use the local repository as a reference repository
when the local repository is a shallow clone.  A shallow clone
repository has no content to offer as a reference.

Fixes 64 test failures when the test repository is a shallow clone.

https://github.com/jenkinsci/bom/pull/1613 describes the experiment to
run the plugin compatibility tests with shallow clones.

https://github.com/jenkinsci/git-plugin/pull/1365 implemented a similar
change in the git plugin so that shallow cloned repositories will not
fail the tests in that plugin.

Will make a few tests in a shallow repository a little slower because
they will require a full clone of the git client plugin repository
rather than reusing the existing clone in the test workspace.  Doubtful
that the performance difference will be detectable within the general
variability of CI test jobs.