Exclude osgi, sftp, and eddsa transitive dependencies
Jenkins does not use OSGi, should not need the osgi jar file and should
not include it in the hpi file because it will clutter the hpi file with
unnecessary content.

The git client plugin has no known use case that requires sftp.  Git does
not use the ftp or sftp protocol.

The Java classes provided by the eddsa library are not referenced anywhere
in the JGit source code.  Interactive testing with ED-25519 keys and JGit
will be needed to confirm that this exclusion does not break relevant
ssh protocol use cases with ED-25519 keys.