Simplify dependabot configuration
Exclusions are no longer needed because versions of the excluded items
are now managed by the plugin bill of materials.

Open pull request limit is not needed because pull requests are merged
or closed promptly.

Target branch is not needed because the default matches this repository.

Reviewers is not needed because I'm already reviewing them regularly
without being assigned as a reviewer.