Simplify loose objects test for CLI git 2.41 (#1019)
Command line git 2.41 adds a new file suffix when it writes a loose
object.  The new file suffix is not relevant to the assertion in this
test.  Simplify the assertion to check that there is at least one
file with the '.idx' suffix and at least one file with the '.pack'
suffix.  Other files are allowed and other suffixes are allowed.