Use proxy.getSecretPassword(), not deprecated proxy.getPassword() (#1072)
Thanks to @StefanSpieker for the pointer to
https://ci.jenkins.io/job/Reporting/job/deprecated-usage-in-plugins/
where it shows that proxy.getPassword() is the only deprecated core API
used in the git client plugin.  Easy to fix.

Not tested by automation.

Tested interactively by temporarily adding a logging statement to report
the value of userInfo after the assignment.  I defined a proxy in the
"Advanced settings" of the plugin manager page, then attempted to clone a
git repository over https.  The logging statement displayed the expected
value in the userInfo string.  I removed the temporary logging statement.