Remove local repo config changes at end of tests
Local configuration changes in GitClientTest allow the tests to run in
environments where commit signing is the user or system level default.
Remove those local configuration changes at the end of the tests, since
they were specifically performed for the test.

Separate the add method from other methods

Easier to understand when BeforeClass has a matching AfterClass

Add more details to the Javadoc so that others understand that config
with a null value will remove the key from the configuration.