Remove last references to JUnit 3 (#1146)
JUnit 3 usage was removed from the git client plugin test suite a year
or two ago.  No need to mention it in comments.

Use the same test skipping test as is used in the git plugin.  The
warnings plugin reports a warning for each test that is not executed
due to an assumeThat() condition.  I don't consider these cases worthy
of a warning because it is intentional that the test is not executed.