Use JGit 7.0.0 with Java 17 (#1170)
JGit 7.0.0 requires Java 17.  Jenkins 2.463 requires Java 17.

Use the plugin bill of materials from 2.462.x because it is the closest
we have to 2.463.  The plugin is expected to work with any release 2.463
or later.

Adapt to removed API's in JGit 7.0.0 by replacing calls to removed API's
with the recommended equivalents.

Tested in my development environment and no issues detected.  Tests run
from a container based Jenkins controller with Windows, Linux, and FreeBSD
agents.  Tests ran on multiple weekly releases of Jenkins since 2.463.
The Linux agents include:

* Alpine
* Debian Linux 11, 12, testing, and unstable
* openSUSE
* Red Hat Enterprise Linux 8
* Rocky Linux 9
* Ubuntu 20.04, 22.04, and 24.04