Minor improvements to JGitLightweightTagTest
Prepare for JGit 7.2.0 where the `packRefs()` method will be publicly
available instead of requiring a JGit internal API.

Document that the packRefs() call is required in order to confirm that
the previous bug in JGitAPIImpl is still fixed.  I reverted the fix and
confirmed that without the call to `packRefs()` the bug is not shown by
the test.

Assert that the set of tags has the correct size.