[JENKINS-73923] Add javadoc for SSH host key verification strategies (#1237)
https://issues.jenkins.io/browse/JENKINS-73923 notes that
https://jenkins.io/doc/developer/extensions/git-client/ has
no documentation.  It says "This extension point has no Javadoc
documentation".

Provide the documentation for the extension point and the four
implementations of the extension point so that extension point
documentation will include documentation after the next release of the
git client plugin.

Much of the text is copied from the README or derived from the text in
the README.

Testing done

Checked that the Javadoc renders as expected in Chrome.