Skip HEAD symref returned by CLI git 2.48.0 and later
Command line git 2.48.0 and later are now able to update the HEAD ref
that is kept locally to represent the condition of the default branch on
the remote repository.  That update was read by the getRemoteBranches()
method as though it were a distinct remote branch rather than a symbolic
reference to the default branch.

Adapt the getRemoteBranches() method to not return a remote branch "HEAD"
if that remote branch is also a symbolic reference.

https://github.blog/open-source/git/highlights-from-git-2-48/ says:

With Git 2.48, if the remote has a default branch but
refs/remotes/origin/HEAD is missing locally, then a fetch will update it.

Testing done:

* Confirmed that command line git 2.48.0 fails testListRemoteBranches
  without this change and passes with this change

* Confirmed that command line git 2.43.0 passes testListRemoteBranches
  without this change and passes with this change